Revert "Sfh fix tests (#8)" (#9)
This reverts commit 5776aad0f1b35b6fdce1bca9407bcda4ed73e17f.

+++ src/coreclr/pal/inc/pal_endian.h
@@ -16,7 +16,7 @@
 extern "C++" {
 inline UINT16 SWAP16(UINT16 x)
 {
-    return (UINT16)((x >> 8) | (x << 8));
+    return (x >> 8) | (x << 8);
 }
 
 inline UINT32 SWAP32(UINT32 x)
