+++ src/mono/mono/utils/monobitset.h
@@ -34,11 +34,11 @@ enum {
 	do \
 	{ \
 		MonoBitSet *__set = (set); \
-		for (gsize __i = 0; __i < __set->size / MONO_BITSET_BITS_PER_CHUNK; __i++) { \
+		for (int __i = 0; __i < __set->size / MONO_BITSET_BITS_PER_CHUNK; __i++) { \
 			if (__set->data [__i]) { \
-				for (gsize __j = 0; __j < MONO_BITSET_BITS_PER_CHUNK; __j++) { \
+				for (int __j = 0; __j < MONO_BITSET_BITS_PER_CHUNK; __j++) { \
 					if (__set->data [__i] & ((gsize) 1 << __j)) { \
-						gsize idx = __j + __i * MONO_BITSET_BITS_PER_CHUNK; \
+						guint idx = __j + __i * MONO_BITSET_BITS_PER_CHUNK; \
 						__VA_ARGS__; \
 					} \
 				} \
